refactor(T5 U1): impl-split SmartSettingsFormWidget across 3 .cpp [BUILD-PASSED]
Slice U1. Pure impl-split (one UWidget class, 3 .cpp); the .h is BYTE-IDENTICAL (BindWidget members
+ UFUNCTIONs unchanged), so the Smart_SettingsForm_Widget BP contract is preserved. _Presets.cpp
(1,259: preset apply/save/import + counter populate + apply/reset + mouse + recipe) + _Controls.cpp
(1,346: belt/pipe/power tier controls + grid + spinbox + confirmation dialogs). SmartSettingsFormWidget.cpp
3,746 -> 1,221. Each split TU re-declares #define/#undef LOCTEXT_NAMESPACE "SmartFoundations" (the
LOCTEXT macros are TU-scoped). Shipping PackagePlugin clean (ExitCode=0). Solo lane; recompile the
widget BP at next deploy (no-op - .h unchanged).
