+++ Source/SmartFoundations/Private/UI/SmartSettingsFormWidget.cpp
@@ -1220,2527 +1220,4 @@ void USmartSettingsFormWidget::CancelAndClose()
     CloseForm();
 }
 
-void USmartSettingsFormWidget::OnApplyPresetClicked()
-{
-    if (!CachedSubsystem.IsValid() || !PresetDropdown)
-    {
-        return;
-    }
-
-    USFRestoreService* RestoreSvc = CachedSubsystem->GetRestoreService();
-    if (!RestoreSvc)
-    {
-        return;
-    }
-
-    const FString SelectedName = PresetDropdown->GetSelectedOption();
-    if (SelectedName.IsEmpty())
-    {
-        return;
-    }
-
-    bool bFound = false;
-    const FSFRestorePreset Preset = RestoreSvc->LoadPreset(SelectedName, bFound);
-    if (!bFound)
-    {
-        return;
-    }
-
-    if (RestoreSvc->ApplyPreset(Preset))
-    {
-        const bool bWasApplyImmediately = bApplyImmediately;
-        const bool bWasApplyImmediatelyEnabled = ApplyImmediatelyCheckBox ? ApplyImmediatelyCheckBox->GetIsEnabled() : true;
-        PopulateFromCounterState(CachedSubsystem.Get());
-        bApplyImmediately = bWasApplyImmediately;
-        if (ApplyImmediatelyCheckBox)
-        {
-            ApplyImmediatelyCheckBox->SetIsChecked(bApplyImmediately);
-            ApplyImmediatelyCheckBox->SetIsEnabled(bWasApplyImmediatelyEnabled);
-        }
-        CacheCurrentStateAsApplied();
-    }
-}
-
-void USmartSettingsFormWidget::OnSavePresetClicked()
-{
-    if (!CachedSubsystem.IsValid() || !PresetNameInput)
-    {
-        return;
-    }
-
-    USFRestoreService* RestoreSvc = CachedSubsystem->GetRestoreService();
-    if (!RestoreSvc)
-    {
-        return;
-    }
-
-    const FString Name = PresetNameInput->GetText().ToString().TrimStartAndEnd();
-    if (Name.IsEmpty())
-    {
-        return;
-    }
-
-    if (RestoreSvc->PresetExists(Name))
-    {
-        PendingConfirmCallback = [this, Name]()
-        {
-            USFRestoreService* Svc = CachedSubsystem.IsValid() ? CachedSubsystem->GetRestoreService() : nullptr;
-            if (!Svc)
-            {
-                return;
-            }
-
-            FSFRestorePreset Preset = Svc->CaptureCurrentState(Name, GetCaptureFlags());
-            Preset.Description = GetPresetDescriptionText();
-            if (Svc->SavePreset(Preset))
-            {
-                RefreshPresetDropdown(Name);
-            }
-        };
-        bWaitingForConfirmation = true;
-        ShowConfirmationDialog(
-            LOCTEXT("Panel_Restore_OverwriteTitle", "Overwrite Preset").ToString(),
-            FText::Format(LOCTEXT("Panel_Restore_OverwriteMessage", "A preset named '{0}' already exists. Overwrite?"), FText::FromString(Name)).ToString(),
-            FLinearColor(1.0f, 0.6f, 0.0f, 1.0f));
-        return;
-    }
-
-    FSFRestorePreset Preset = RestoreSvc->CaptureCurrentState(Name, GetCaptureFlags());
-    Preset.Description = GetPresetDescriptionText();
-    if (RestoreSvc->SavePreset(Preset))
-    {
-        RefreshPresetDropdown(Name);
-    }
-}
-
-void USmartSettingsFormWidget::OnDeletePresetClicked()
-{
-    if (!CachedSubsystem.IsValid() || !PresetDropdown)
-    {
-        return;
-    }
-
-    const FString SelectedName = PresetDropdown->GetSelectedOption();
-    if (SelectedName.IsEmpty())
-    {
-        return;
-    }
-
-    PendingConfirmCallback = [this, SelectedName]()
-    {
-        USFRestoreService* Svc = CachedSubsystem.IsValid() ? CachedSubsystem->GetRestoreService() : nullptr;
-        if (!Svc)
-        {
-            return;
-        }
-
-        if (Svc->DeletePreset(SelectedName))
-        {
-            RefreshPresetDropdown();
-        }
-    };
-    bWaitingForConfirmation = true;
-    ShowConfirmationDialog(
-        LOCTEXT("Panel_Restore_DeleteTitle", "Delete Preset").ToString(),
-        FText::Format(LOCTEXT("Panel_Restore_DeleteMessage", "Delete preset '{0}'?"), FText::FromString(SelectedName)).ToString(),
-        FLinearColor(1.0f, 0.3f, 0.3f, 1.0f));
-}
-
-void USmartSettingsFormWidget::OnUpdatePresetClicked()
-{
-    if (!CachedSubsystem.IsValid() || !PresetDropdown)
-    {
-        return;
-    }
-
-    USFRestoreService* RestoreSvc = CachedSubsystem->GetRestoreService();
-    if (!RestoreSvc)
-    {
-        return;
-    }
-
-    const FString SelectedName = PresetDropdown->GetSelectedOption();
-    if (SelectedName.IsEmpty())
-    {
-        return;
-    }
-
-    FSFRestorePreset Preset = RestoreSvc->CaptureCurrentState(SelectedName, GetCaptureFlags());
-    Preset.Description = GetPresetDescriptionText();
-    if (RestoreSvc->SavePreset(Preset))
-    {
-        RefreshPresetDropdown(SelectedName);
-    }
-}
-
-void USmartSettingsFormWidget::OnExportPresetClicked()
-{
-    if (!CachedSubsystem.IsValid() || !PresetDropdown)
-    {
-        return;
-    }
-
-    USFRestoreService* RestoreSvc = CachedSubsystem->GetRestoreService();
-    if (!RestoreSvc)
-    {
-        return;
-    }
-
-    const FString SelectedName = PresetDropdown->GetSelectedOption();
-    if (SelectedName.IsEmpty())
-    {
-        return;
-    }
-
-    bool bFound = false;
-    const FSFRestorePreset Preset = RestoreSvc->LoadPreset(SelectedName, bFound);
-    if (!bFound)
-    {
-        return;
-    }
-
-    const FString Encoded = RestoreSvc->ExportToString(Preset);
-    FPlatformApplicationMisc::ClipboardCopy(*Encoded);
-}
-
-void USmartSettingsFormWidget::OnImportPresetClicked()
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    USFRestoreService* RestoreSvc = CachedSubsystem->GetRestoreService();
-    if (!RestoreSvc)
-    {
-        return;
-    }
-
-    FString ClipboardText;
-    FPlatformApplicationMisc::ClipboardPaste(ClipboardText);
-    if (ClipboardText.IsEmpty())
-    {
-        return;
-    }
-
-    bool bSuccess = false;
-    const FSFRestorePreset Preset = RestoreSvc->ImportFromString(ClipboardText, bSuccess);
-    if (!bSuccess)
-    {
-        return;
-    }
-
-    if (RestoreSvc->SavePreset(Preset))
-    {
-        RefreshPresetDropdown(Preset.Name);
-    }
-}
-
-void USmartSettingsFormWidget::OnImportFromExtendClicked()
-{
-    UpdateExtendImportButtonState();
-
-    if (!CachedSubsystem.IsValid() || !PresetNameInput)
-    {
-        SF_RESTORE_DIAGNOSTIC_LOG(LogSmartFoundations, Warning,
-            TEXT("[SmartRestore][UI] ImportFromExtend clicked but widget state is invalid: CachedSubsystem=%d PresetNameInput=%d"),
-            CachedSubsystem.IsValid() ? 1 : 0,
-            PresetNameInput ? 1 : 0);
-        return;
-    }
-
-    USFRestoreService* RestoreSvc = CachedSubsystem->GetRestoreService();
-    if (!RestoreSvc || !RestoreSvc->IsLastExtendAvailable())
-    {
-        SF_RESTORE_DIAGNOSTIC_LOG(LogSmartFoundations, Warning,
-            TEXT("[SmartRestore][UI] ImportFromExtend unavailable: RestoreSvc=%s LastExtendAvailable=%d"),
-            RestoreSvc ? TEXT("valid") : TEXT("null"),
-            RestoreSvc ? (RestoreSvc->IsLastExtendAvailable() ? 1 : 0) : 0);
-        return;
-    }
-
-    FString Name = PresetNameInput->GetText().ToString().TrimStartAndEnd();
-    if (Name.IsEmpty())
-    {
-        Name = FString::Printf(TEXT("Extend Source %s"), *FDateTime::Now().ToString(TEXT("%Y%m%d-%H%M%S")));
-        PresetNameInput->SetText(FText::FromString(Name));
-    }
-
-    bool bSuccess = false;
-    FSFRestorePreset Preset = RestoreSvc->ImportFromLastExtend(Name, GetCaptureFlags(), bSuccess);
-    if (!bSuccess)
-    {
-        SF_RESTORE_DIAGNOSTIC_LOG(LogSmartFoundations, Warning,
-            TEXT("[SmartRestore][UI] ImportFromLastExtend failed for preset '%s'"),
-            *Name);
-        return;
-    }
-
-    Preset.CaptureFlags.bGrid = true;
-    Preset.GridCounters = FIntVector(1, 1, 1);
-    Preset.Description = GetPresetDescriptionText();
-
-    const bool bApplied = RestoreSvc->ApplyPreset(Preset);
-    SF_RESTORE_DIAGNOSTIC_LOG(LogSmartFoundations, Log,
-        TEXT("[SmartRestore][UI] ImportFromExtend staged for editing: preset='%s' applied=%d hasTopology=%d childHolograms=%d grid=(%d,%d,%d)"),
-        *Name,
-        bApplied ? 1 : 0,
-        Preset.bHasExtendTopology ? 1 : 0,
-        Preset.ExtendCloneTopology.ChildHolograms.Num(),
-        Preset.GridCounters.X,
-        Preset.GridCounters.Y,
-        Preset.GridCounters.Z);
-
-    if (bApplied)
-    {
-        const bool bWasApplyImmediately = bApplyImmediately;
-        const bool bWasApplyImmediatelyEnabled = ApplyImmediatelyCheckBox ? ApplyImmediatelyCheckBox->GetIsEnabled() : true;
-        PopulateFromCounterState(CachedSubsystem.Get());
-        bApplyImmediately = bWasApplyImmediately;
-        if (ApplyImmediatelyCheckBox)
-        {
-            ApplyImmediatelyCheckBox->SetIsChecked(bApplyImmediately);
-            ApplyImmediatelyCheckBox->SetIsEnabled(bWasApplyImmediatelyEnabled);
-        }
-        CacheCurrentStateAsApplied();
-    }
-}
-
-void USmartSettingsFormWidget::OnPresetSelectionChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    UpdateRestorePresetDetails(SelectedItem);
-
-    if (SelectionType == ESelectInfo::Direct || !CachedSubsystem.IsValid() || SelectedItem.IsEmpty())
-    {
-        return;
-    }
-
-    USFRestoreService* RestoreSvc = CachedSubsystem->GetRestoreService();
-    if (!RestoreSvc)
-    {
-        return;
-    }
-
-    bool bFound = false;
-    const FSFRestorePreset Preset = RestoreSvc->LoadPreset(SelectedItem, bFound);
-    if (bFound)
-    {
-        PopulateSmartPanelFromPreset(Preset);
-    }
-}
-
-void USmartSettingsFormWidget::OnRestoreSectionToggleClicked()
-{
-    UWidget* RestorePanel = RestoreSidePanel ? Cast<UWidget>(RestoreSidePanel) : Cast<UWidget>(RestoreContainer);
-    if (!RestorePanel)
-    {
-        return;
-    }
-
-    const bool bVisible = RestorePanel->GetVisibility() != ESlateVisibility::Collapsed;
-    RestorePanel->SetVisibility(bVisible ? ESlateVisibility::Collapsed : ESlateVisibility::Visible);
-}
-
-void USmartSettingsFormWidget::RefreshPresetDropdown(const FString& PreferredSelection)
-{
-    if (!PresetDropdown || !CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    USFRestoreService* RestoreSvc = CachedSubsystem->GetRestoreService();
-    if (!RestoreSvc)
-    {
-        return;
-    }
-
-    const FString PreviousSelection = PresetDropdown->GetSelectedOption();
-    PresetDropdown->ClearOptions();
-
-    const TArray<FString> Names = RestoreSvc->GetPresetNames();
-    for (const FString& Name : Names)
-    {
-        PresetDropdown->AddOption(Name);
-    }
-
-    FString SelectionToApply;
-    if (!PreferredSelection.IsEmpty() && Names.Contains(PreferredSelection))
-    {
-        SelectionToApply = PreferredSelection;
-    }
-    else if (!PreviousSelection.IsEmpty() && Names.Contains(PreviousSelection))
-    {
-        SelectionToApply = PreviousSelection;
-    }
-
-    if (!SelectionToApply.IsEmpty())
-    {
-        PresetDropdown->SetSelectedOption(SelectionToApply);
-        UpdateRestorePresetDetails(SelectionToApply);
-    }
-    else if (Names.Num() > 0)
-    {
-        PresetDropdown->SetSelectedIndex(0);
-        UpdateRestorePresetDetails(Names[0]);
-    }
-    else
-    {
-        UpdateRestorePresetDetails(FString());
-    }
-}
-
-void USmartSettingsFormWidget::UpdateRestorePresetDetails(const FString& PresetName)
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    if (PresetName.IsEmpty())
-    {
-        if (PresetDescriptionInput)
-        {
-            PresetDescriptionInput->SetText(FText::GetEmpty());
-        }
-        if (PresetCreatedAtValue)
-        {
-            PresetCreatedAtValue->SetText(LOCTEXT("Panel_Restore_NotSaved", "Not saved"));
-        }
-        return;
-    }
-
-    USFRestoreService* RestoreSvc = CachedSubsystem->GetRestoreService();
-    if (!RestoreSvc)
-    {
-        return;
-    }
-
-    bool bFound = false;
-    const FSFRestorePreset Preset = RestoreSvc->LoadPreset(PresetName, bFound);
-    if (!bFound)
-    {
-        return;
-    }
-
-    if (PresetDescriptionInput)
-    {
-        PresetDescriptionInput->SetText(FText::FromString(Preset.Description));
-    }
-    if (PresetCreatedAtValue)
-    {
-        PresetCreatedAtValue->SetText(FText::FromString(FormatPresetTimestampForDisplay(Preset.CreatedAt)));
-    }
-}
-
-void USmartSettingsFormWidget::PopulateSmartPanelFromPreset(const FSFRestorePreset& Preset)
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    const FSFCounterState PendingState = BuildPendingCounterStateFromPreset(Preset);
-    PopulateCounterInputsFromState(PendingState);
-    SF_RESTORE_DIAGNOSTIC_LOG(LogSmartFoundations, Log,
-        TEXT("[SmartRestore][UI] Loaded preset '%s' into Smart Panel pending values: grid=%d spacing=%d steps=%d stagger=%d rotation=%d"),
-        *Preset.Name,
-        Preset.CaptureFlags.bGrid ? 1 : 0,
-        Preset.CaptureFlags.bSpacing ? 1 : 0,
-        Preset.CaptureFlags.bSteps ? 1 : 0,
-        Preset.CaptureFlags.bStagger ? 1 : 0,
-        Preset.CaptureFlags.bRotation ? 1 : 0);
-}
-
-FSFCounterState USmartSettingsFormWidget::BuildPendingCounterStateFromPreset(const FSFRestorePreset& Preset) const
-{
-    FSFCounterState State = CachedSubsystem.IsValid()
-        ? CachedSubsystem->GetCounterState()
-        : FSFCounterState();
-
-    if (Preset.CaptureFlags.bGrid)
-    {
-        State.GridCounters = Preset.GridCounters;
-    }
-    if (Preset.CaptureFlags.bSpacing)
-    {
-        State.SpacingX = Preset.SpacingX;
-        State.SpacingY = Preset.SpacingY;
-        State.SpacingZ = Preset.SpacingZ;
-    }
-    if (Preset.CaptureFlags.bSteps)
-    {
-        State.StepsX = Preset.StepsX;
-        State.StepsY = Preset.StepsY;
-    }
-    if (Preset.CaptureFlags.bStagger)
-    {
-        State.StaggerX = Preset.StaggerX;
-        State.StaggerY = Preset.StaggerY;
-        State.StaggerZX = Preset.StaggerZX;
-        State.StaggerZY = Preset.StaggerZY;
-    }
-    if (Preset.CaptureFlags.bRotation)
-    {
-        State.RotationZ = Preset.RotationZ;
-    }
-
-    return State;
-}
-
-void USmartSettingsFormWidget::PopulateCounterInputsFromState(const FSFCounterState& State)
-{
-    const bool bWasApplyImmediately = bApplyImmediately;
-    bApplyImmediately = false;
-
-    bGridXPositive = State.GridCounters.X >= 0;
-    bGridYPositive = State.GridCounters.Y >= 0;
-    bGridZPositive = State.GridCounters.Z >= 0;
-
-    if (GridXInput)
-    {
-        GridXInput->SetValue(static_cast<float>(FMath::Abs(State.GridCounters.X)));
-    }
-    if (GridYInput)
-    {
-        GridYInput->SetValue(static_cast<float>(FMath::Abs(State.GridCounters.Y)));
-    }
-    if (GridZInput)
-    {
-        GridZInput->SetValue(static_cast<float>(FMath::Abs(State.GridCounters.Z)));
-    }
-
-    UpdateGridDirectionLabel(GridXDirLabel, bGridXPositive);
-    UpdateGridDirectionLabel(GridYDirLabel, bGridYPositive);
-    UpdateGridDirectionLabel(GridZDirLabel, bGridZPositive);
-
-    if (SpacingXInput)
-    {
-        SpacingXInput->SetValue(State.SpacingX / 100.0f);
-    }
-    if (SpacingYInput)
-    {
-        SpacingYInput->SetValue(State.SpacingY / 100.0f);
-    }
-    if (SpacingZInput)
-    {
-        SpacingZInput->SetValue(State.SpacingZ / 100.0f);
-    }
-
-    if (StepsXInput)
-    {
-        StepsXInput->SetValue(State.StepsX / 100.0f);
-    }
-    if (StepsYInput)
-    {
-        StepsYInput->SetValue(State.StepsY / 100.0f);
-    }
-
-    if (StaggerXInput)
-    {
-        StaggerXInput->SetValue(State.StaggerX / 100.0f);
-    }
-    if (StaggerYInput)
-    {
-        StaggerYInput->SetValue(State.StaggerY / 100.0f);
-    }
-    if (StaggerZXInput)
-    {
-        StaggerZXInput->SetValue(State.StaggerZX / 100.0f);
-    }
-    if (StaggerZYInput)
-    {
-        StaggerZYInput->SetValue(State.StaggerZY / 100.0f);
-    }
-
-    if (RotationZInput)
-    {
-        RotationZInput->SetValue(State.RotationZ);
-    }
-
-    bApplyImmediately = bWasApplyImmediately;
-    UpdateGridWarningDisplay();
-}
-
-FString USmartSettingsFormWidget::GetPresetDescriptionText() const
-{
-    return PresetDescriptionInput
-        ? PresetDescriptionInput->GetText().ToString().TrimStartAndEnd()
-        : FString();
-}
-
-FString USmartSettingsFormWidget::FormatPresetTimestampForDisplay(const FString& IsoTimestamp) const
-{
-    if (IsoTimestamp.IsEmpty())
-    {
-        return LOCTEXT("Panel_Restore_NotSaved", "Not saved").ToString();
-    }
-
-    FDateTime UtcTimestamp;
-    if (!FDateTime::ParseIso8601(*IsoTimestamp, UtcTimestamp))
-    {
-        return IsoTimestamp;
-    }
-
-    const FTimespan LocalOffset = FDateTime::Now() - FDateTime::UtcNow();
-    const FDateTime LocalTimestamp = UtcTimestamp + LocalOffset;
-    return LocalTimestamp.ToString(TEXT("%Y-%m-%d %H:%M Local"));
-}
-
-FSFRestoreCaptureFlags USmartSettingsFormWidget::GetCaptureFlags() const
-{
-    FSFRestoreCaptureFlags Flags;
-    Flags.bGrid = CaptureGridCheckBox ? CaptureGridCheckBox->IsChecked() : true;
-    Flags.bSpacing = CaptureSpacingCheckBox ? CaptureSpacingCheckBox->IsChecked() : true;
-    Flags.bSteps = CaptureStepsCheckBox ? CaptureStepsCheckBox->IsChecked() : true;
-    Flags.bStagger = CaptureStaggerCheckBox ? CaptureStaggerCheckBox->IsChecked() : true;
-    Flags.bRotation = CaptureRotationCheckBox ? CaptureRotationCheckBox->IsChecked() : true;
-    Flags.bRecipe = CaptureRecipeCheckBox ? CaptureRecipeCheckBox->IsChecked() : true;
-    Flags.bAutoConnect = CaptureAutoConnectCheckBox ? CaptureAutoConnectCheckBox->IsChecked() : true;
-    return Flags;
-}
-
-void USmartSettingsFormWidget::UpdateExtendImportButtonState()
-{
-    if (!ImportFromExtendBtn)
-    {
-        return;
-    }
-
-    USFRestoreService* RestoreSvc = CachedSubsystem.IsValid() ? CachedSubsystem->GetRestoreService() : nullptr;
-    ImportFromExtendBtn->SetIsEnabled(RestoreSvc && RestoreSvc->IsLastExtendAvailable());
-    UpdateRestoreButtonTextColors();
-}
-
-void USmartSettingsFormWidget::UpdateRestoreButtonTextColors()
-{
-    const FSlateColor BlackText(FLinearColor::Black);
-    const FSlateColor DisabledImportText(FLinearColor(0.55f, 0.55f, 0.55f, 1.0f));
-
-    auto SetTextColor = [this](const TCHAR* WidgetName, const FSlateColor& Color)
-    {
-        if (UTextBlock* TextBlock = Cast<UTextBlock>(GetWidgetFromName(FName(WidgetName))))
-        {
-            TextBlock->SetColorAndOpacity(Color);
-        }
-    };
-
-    SetTextColor(TEXT("RestoreSectionHeader"), BlackText);
-    SetTextColor(TEXT("ApplyPresetBtnText"), BlackText);
-    SetTextColor(TEXT("SavePresetBtnText"), BlackText);
-    SetTextColor(TEXT("DeletePresetBtnText"), BlackText);
-    SetTextColor(TEXT("UpdatePresetBtnText"), BlackText);
-    SetTextColor(TEXT("ExportPresetBtnText"), BlackText);
-    SetTextColor(TEXT("ImportPresetBtnText"), BlackText);
-    SetTextColor(TEXT("ImportFromExtendBtnText"),
-        ImportFromExtendBtn && ImportFromExtendBtn->GetIsEnabled() ? BlackText : DisabledImportText);
-}
-
-void USmartSettingsFormWidget::CloseForm()
-{
-    // Restore HUD visibility
-    if (CachedSubsystem.IsValid())
-    {
-        if (USFHudService* HudService = CachedSubsystem->GetHudService())
-        {
-            HudService->SetHUDSuppressed(false);
-        }
-    }
-
-    // Restore game input mode and hide cursor
-    APlayerController* PC = GetOwningPlayer();
-    if (PC)
-    {
-        PC->bShowMouseCursor = false;
-        FInputModeGameOnly InputMode;
-        PC->SetInputMode(InputMode);
-    }
-
-    // Remove widget from viewport
-    RemoveFromParent();
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Closed and input restored"));
-}
-
-void USmartSettingsFormWidget::OnApplyButtonClicked()
-{
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Apply button clicked"));
-
-    // Don't process if we're already waiting for confirmation
-    if (bWaitingForConfirmation)
-    {
-        UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Already waiting for confirmation, ignoring apply"));
-        return;
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        UE_LOG(LogSmartFoundations, Error, TEXT("Settings Form: No cached subsystem reference"));
-        return;
-    }
-
-    // Check if confirmation is required for large grids
-    if (RequiresApplyConfirmation())
-    {
-        // Show confirmation dialog and defer the actual apply
-        ShowLargeGridConfirmation([this]()
-        {
-            // This lambda is called when user confirms
-            ApplyCurrentValues();
-        });
-        return;
-    }
-
-    // No confirmation needed, apply directly
-    ApplyCurrentValues();
-}
-
-void USmartSettingsFormWidget::OnResetButtonClicked()
-{
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Reset button clicked - zeroing spacing, steps, stagger, and rotation"));
-
-    // Zero out spacing SpinBoxes
-    if (SpacingXInput) SpacingXInput->SetValue(0.0f);
-    if (SpacingYInput) SpacingYInput->SetValue(0.0f);
-    if (SpacingZInput) SpacingZInput->SetValue(0.0f);
-
-    // Zero out steps SpinBoxes
-    if (StepsXInput) StepsXInput->SetValue(0.0f);
-    if (StepsYInput) StepsYInput->SetValue(0.0f);
-
-    // Zero out stagger SpinBoxes
-    if (StaggerXInput) StaggerXInput->SetValue(0.0f);
-    if (StaggerYInput) StaggerYInput->SetValue(0.0f);
-    if (StaggerZXInput) StaggerZXInput->SetValue(0.0f);
-    if (StaggerZYInput) StaggerZYInput->SetValue(0.0f);
-
-    // Zero out rotation SpinBox
-    if (RotationZInput) RotationZInput->SetValue(0.0f);
-
-    // Apply the reset values immediately (regardless of Apply Immediately mode)
-    OnApplyButtonClicked();
-}
-
-void USmartSettingsFormWidget::ApplyCurrentValues()
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        UE_LOG(LogSmartFoundations, Error, TEXT("Settings Form: No cached subsystem reference"));
-        return;
-    }
-
-    // Parse values from editable text inputs
-    FSFCounterState NewState = CachedSubsystem->GetCounterState();
-
-    // Read Grid values from SpinBox (absolute count, apply direction from toggle state)
-    if (GridXInput)
-    {
-        int32 Count = FMath::Max(1, FMath::RoundToInt(GridXInput->GetValue()));
-        NewState.GridCounters.X = bGridXPositive ? Count : -Count;
-    }
-    if (GridYInput)
-    {
-        int32 Count = FMath::Max(1, FMath::RoundToInt(GridYInput->GetValue()));
-        NewState.GridCounters.Y = bGridYPositive ? Count : -Count;
-    }
-    if (GridZInput)
-    {
-        int32 Count = FMath::Max(1, FMath::RoundToInt(GridZInput->GetValue()));
-        NewState.GridCounters.Z = bGridZPositive ? Count : -Count;
-    }
-
-    // Read Spacing values from SpinBox (meters in UI -> centimeters in state)
-    if (SpacingXInput)
-    {
-        NewState.SpacingX = FMath::RoundToInt(SpacingXInput->GetValue() * 100.0f);
-    }
-    if (SpacingYInput)
-    {
-        NewState.SpacingY = FMath::RoundToInt(SpacingYInput->GetValue() * 100.0f);
-    }
-    if (SpacingZInput)
-    {
-        NewState.SpacingZ = FMath::RoundToInt(SpacingZInput->GetValue() * 100.0f);
-    }
-
-    // Read Steps values from SpinBox (meters in UI -> centimeters in state)
-    if (StepsXInput)
-    {
-        NewState.StepsX = FMath::RoundToInt(StepsXInput->GetValue() * 100.0f);
-    }
-    if (StepsYInput)
-    {
-        NewState.StepsY = FMath::RoundToInt(StepsYInput->GetValue() * 100.0f);
-    }
-
-    // Read Stagger values from SpinBox (meters in UI -> centimeters in state)
-    if (StaggerXInput)
-    {
-        NewState.StaggerX = FMath::RoundToInt(StaggerXInput->GetValue() * 100.0f);
-    }
-    if (StaggerYInput)
-    {
-        NewState.StaggerY = FMath::RoundToInt(StaggerYInput->GetValue() * 100.0f);
-    }
-    if (StaggerZXInput)
-    {
-        NewState.StaggerZX = FMath::RoundToInt(StaggerZXInput->GetValue() * 100.0f);
-    }
-    if (StaggerZYInput)
-    {
-        NewState.StaggerZY = FMath::RoundToInt(StaggerZYInput->GetValue() * 100.0f);
-    }
-
-    // Read Rotation value from SpinBox (degrees - no conversion needed)
-    if (RotationZInput)
-    {
-        NewState.RotationZ = RotationZInput->GetValue();
-    }
-
-    // Log parsed values before applying
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Parsed values - Grid[%d,%d,%d] Spacing[%d,%d,%d] Steps[%d,%d] Stagger[%d,%d,%d,%d] Rotation[%.1f]"),
-        NewState.GridCounters.X, NewState.GridCounters.Y, NewState.GridCounters.Z,
-        NewState.SpacingX, NewState.SpacingY, NewState.SpacingZ,
-        NewState.StepsX, NewState.StepsY,
-        NewState.StaggerX, NewState.StaggerY, NewState.StaggerZX, NewState.StaggerZY,
-        NewState.RotationZ);
-
-    // Apply the new state (updates GridStateService, local CounterState, and triggers HUD refresh)
-    CachedSubsystem->UpdateCounterState(NewState);
-
-    // Regenerate child holograms to reflect new grid counts and reposition with new spacing/steps/stagger
-    // During Extend mode, UpdateCounterState already triggers OnScaledExtendStateChanged which handles regeneration
-    if (!CachedSubsystem->IsExtendModeActive())
-    {
-        CachedSubsystem->RegenerateChildHologramGrid();
-    }
-
-    // Refresh recipe details to update grid totals with new grid size
-    if (RecipeDetailsContainer && RecipeComboBox)
-    {
-        int32 SelectedIndex = RecipeComboBox->GetSelectedIndex();
-        // Index 0 is "None Selected", actual recipes start at index 1
-        if (SelectedIndex > 0 && SelectedIndex < CachedRecipeList.Num())
-        {
-            TSubclassOf<UFGRecipe> SelectedRecipe = CachedRecipeList[SelectedIndex];
-            if (SelectedRecipe)
-            {
-                PopulateRecipeDetails(SelectedRecipe);
-            }
-        }
-    }
-
-    // Cache the applied state for cancel/revert
-    CacheCurrentStateAsApplied();
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: State applied and grid regenerated"));
-}
-
-FReply USmartSettingsFormWidget::NativeOnMouseButtonDown(const FGeometry& InGeometry, const FPointerEvent& InMouseEvent)
-{
-    if (InMouseEvent.GetEffectingButton() == EKeys::RightMouseButton)
-    {
-        const FVector2D LocalMouse = InGeometry.AbsoluteToLocal(InMouseEvent.GetScreenSpacePosition());
-
-        if (RestoreSidePanelSlot && RestoreSidePanel && RestoreSidePanel->GetVisibility() != ESlateVisibility::Collapsed)
-        {
-            const FVector2D RestorePos = RestoreSidePanelSlot->GetPosition();
-            const FVector2D RestoreSize = RestoreSidePanelSlot->GetSize();
-            const bool bMouseOverRestorePanel =
-                LocalMouse.X >= RestorePos.X && LocalMouse.X <= RestorePos.X + RestoreSize.X &&
-                LocalMouse.Y >= RestorePos.Y && LocalMouse.Y <= RestorePos.Y + RestoreSize.Y;
-
-            if (bMouseOverRestorePanel)
-            {
-                bIsDraggingRestorePanel = true;
-                DragOffset = LocalMouse - RestorePos;
-                return FReply::Handled();
-            }
-        }
-
-        if (BackgroundPanelSlot)
-        {
-            bIsDragging = true;
-            DragOffset = LocalMouse - BackgroundPanelSlot->GetPosition();
-            return FReply::Handled();
-        }
-    }
-
-    return Super::NativeOnMouseButtonDown(InGeometry, InMouseEvent);
-}
-
-FReply USmartSettingsFormWidget::NativeOnMouseButtonUp(const FGeometry& InGeometry, const FPointerEvent& InMouseEvent)
-{
-    if (InMouseEvent.GetEffectingButton() == EKeys::RightMouseButton)
-    {
-        bIsDragging = false;
-        bIsDraggingRestorePanel = false;
-    }
-
-    return Super::NativeOnMouseButtonUp(InGeometry, InMouseEvent);
-}
-
-FReply USmartSettingsFormWidget::NativeOnMouseMove(const FGeometry& InGeometry, const FPointerEvent& InMouseEvent)
-{
-    if (bIsDraggingRestorePanel && RestoreSidePanelSlot)
-    {
-        const FVector2D LocalMouse = InGeometry.AbsoluteToLocal(InMouseEvent.GetScreenSpacePosition());
-        RestoreSidePanelSlot->SetPosition(LocalMouse - DragOffset);
-        return FReply::Handled();
-    }
-
-    if (bIsDragging && BackgroundPanelSlot)
-    {
-        const FVector2D LocalMouse = InGeometry.AbsoluteToLocal(InMouseEvent.GetScreenSpacePosition());
-        const FVector2D NewBackgroundPos = LocalMouse - DragOffset;
-
-        BackgroundPanelSlot->SetPosition(NewBackgroundPos);
-
-        if (TitleTextSlot)
-        {
-            TitleTextSlot->SetPosition(NewBackgroundPos + TitleOffset);
-        }
-        if (ContentContainerSlot)
-        {
-            ContentContainerSlot->SetPosition(NewBackgroundPos + ContentOffset);
-        }
-        if (SmartLogoSlot)
-        {
-            SmartLogoSlot->SetPosition(NewBackgroundPos + LogoOffset);
-        }
-        if (ApplyButtonSlot)
-        {
-            ApplyButtonSlot->SetPosition(NewBackgroundPos + ApplyOffset);
-        }
-        if (ResetButtonSlot)
-        {
-            ResetButtonSlot->SetPosition(NewBackgroundPos + ResetOffset);
-        }
-        if (CloseButtonSlot)
-        {
-            CloseButtonSlot->SetPosition(NewBackgroundPos + CloseOffset);
-        }
-    }
-
-    return Super::NativeOnMouseMove(InGeometry, InMouseEvent);
-}
-void USmartSettingsFormWidget::OnRecipeSelectionChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    // Ignore programmatic selection (when we set it during population)
-    if (SelectionType == ESelectInfo::Direct)
-    {
-        return;
-    }
-
-    if (!CachedSubsystem.IsValid() || !RecipeComboBox)
-    {
-        return;
-    }
-
-    int32 SelectedIndex = RecipeComboBox->GetSelectedIndex();
-    if (SelectedIndex < 0 || SelectedIndex >= CachedRecipeList.Num())
-    {
-        UE_LOG(LogSmartFoundations, Warning, TEXT("Settings Form: Invalid recipe selection index %d"), SelectedIndex);
-        return;
-    }
-
-    TSubclassOf<UFGRecipe> SelectedRecipe = CachedRecipeList[SelectedIndex];
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Recipe selected - Index=%d, Recipe=%s"),
-        SelectedIndex, SelectedRecipe ? *SelectedRecipe->GetName() : TEXT("None"));
-
-    // Check if "None Selected" was chosen (index 0, nullptr recipe)
-    if (SelectedIndex == 0 || !SelectedRecipe)
-    {
-        // Clear the recipe - same as clicking Clear button
-        CachedSubsystem->ClearAllRecipes();
-        CachedSubsystem->ApplyRecipeToParentHologram();
-
-        // Clear the recipe details
-        if (RecipeDetailsContainer)
-        {
-            RecipeDetailsContainer->ClearChildren();
-        }
-
-        // Hide the recipe icon
-        if (RecipeIcon)
-        {
-            RecipeIcon->SetVisibility(ESlateVisibility::Collapsed);
-        }
-
-        UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Recipe cleared via 'None Selected'"));
-        return;
-    }
-
-    // Apply the recipe selection via subsystem
-    // Note: SelectedIndex includes the "None Selected" offset, so subtract 1 for actual recipe index
-    if (USFRecipeManagementService* RecipeService = CachedSubsystem->GetRecipeManagementService())
-    {
-        RecipeService->SetActiveRecipeByIndex(SelectedIndex - 1);
-
-        // Update the recipe details container with icons
-        if (RecipeDetailsContainer)
-        {
-            PopulateRecipeDetails(SelectedRecipe);
-        }
-
-        // Update recipe icon
-        UpdateRecipeIcon(SelectedRecipe);
-    }
-}
-
-void USmartSettingsFormWidget::OnClearRecipeButtonClicked()
-{
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Clear Recipe button clicked"));
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    // Use the existing ClearAllRecipes method on the subsystem
-    // This is the same logic triggered by Num0 when in Recipe mode (U held)
-    CachedSubsystem->ClearAllRecipes();
-
-    // Apply the clear to the parent hologram and trigger child regeneration
-    CachedSubsystem->ApplyRecipeToParentHologram();
-
-    // Clear the ComboBox selection
-    if (RecipeComboBox)
-    {
-        RecipeComboBox->ClearSelection();
-    }
-
-    // Clear the recipe details
-    if (RecipeDetailsContainer)
-    {
-        RecipeDetailsContainer->ClearChildren();
-    }
-
-    // Hide the recipe icon
-    if (RecipeIcon)
-    {
-        RecipeIcon->SetVisibility(ESlateVisibility::Collapsed);
-    }
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Recipe cleared via ClearAllRecipes"));
-}
-
-void USmartSettingsFormWidget::UpdateRecipeIcon(TSubclassOf<UFGRecipe> Recipe)
-{
-    if (!RecipeIcon)
-    {
-        return;
-    }
-
-    if (!Recipe)
-    {
-        RecipeIcon->SetVisibility(ESlateVisibility::Collapsed);
-        return;
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    if (USFRecipeManagementService* RecipeService = CachedSubsystem->GetRecipeManagementService())
-    {
-        UTexture2D* IconTexture = RecipeService->GetRecipePrimaryProductIcon(Recipe);
-        if (IconTexture)
-        {
-            RecipeIcon->SetBrushFromTexture(IconTexture);
-            // Set a fixed square size for the icon (matches ComboBox height)
-            RecipeIcon->SetDesiredSizeOverride(FVector2D(40.0f, 40.0f));
-            RecipeIcon->SetVisibility(ESlateVisibility::Visible);
-        }
-        else
-        {
-            RecipeIcon->SetVisibility(ESlateVisibility::Collapsed);
-        }
-    }
-}
-
-void USmartSettingsFormWidget::PopulateRecipeDetails(TSubclassOf<UFGRecipe> Recipe)
-{
-    if (!RecipeDetailsContainer)
-    {
-        return;
-    }
-
-    // Clear existing children
-    RecipeDetailsContainer->ClearChildren();
-
-    if (!Recipe || !CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    UFGRecipe* RecipeCDO = Recipe->GetDefaultObject<UFGRecipe>();
-    if (!RecipeCDO)
-    {
-        return;
-    }
-
-    // Colors matching HUD style
-    const FLinearColor OutputColor(0.7f, 0.9f, 0.7f, 1.0f);   // Light green for outputs
-    const FLinearColor InputColor(0.9f, 0.8f, 0.7f, 1.0f);    // Light tan for inputs
-
-    // Get manufacturing duration to calculate per-minute rates
-    // Duration is in seconds, so items per minute = (Amount / Duration) * 60
-    const float ManufacturingDuration = UFGRecipe::GetManufacturingDuration(Recipe);
-    const float CyclesPerMinute = (ManufacturingDuration > 0.0f) ? (60.0f / ManufacturingDuration) : 0.0f;
-
-    // Get grid size to calculate total output for the entire grid
-    // Grid total = GridX * GridY * GridZ buildings
-    // Note: Grid values can be negative (direction), so use Abs() to get the count
-    int32 GridTotal = 1;
-    if (CachedSubsystem.IsValid())
-    {
-        FSFCounterState CounterState = CachedSubsystem->GetCounterState();
-        // GridCounters is an FIntVector with X, Y, Z components - use absolute values since sign indicates direction
-        GridTotal = FMath::Max(1, FMath::Abs(CounterState.GridCounters.X)) *
-                    FMath::Max(1, FMath::Abs(CounterState.GridCounters.Y)) *
-                    FMath::Max(1, FMath::Abs(CounterState.GridCounters.Z));
-    }
-
-    // No header row needed - the ComboBox provides recipe selection
-    // Go directly to output/input rows with icons
-    TArray<FItemAmount> Products = RecipeCDO->GetProducts();
-
-    // Add output rows with icons
-    for (int32 i = 0; i < Products.Num(); i++)
-    {
-        TSubclassOf<UFGItemDescriptor> ItemClass = Products[i].ItemClass;
-        int32 Amount = Products[i].Amount;
-        FString ItemName = UFGItemDescriptor::GetItemName(ItemClass).ToString();
-
-        // Get item icon
-        UTexture2D* ItemIcon = UFGItemDescriptor::GetSmallIcon(ItemClass);
-
-        // Calculate per-minute rate (single building) and grid total
-        float PerMinute = Amount * CyclesPerMinute;
-        float GridTotalPerMinute = PerMinute * GridTotal;
-
-        // Format amount with per-minute rate and grid total (liquids have amounts >= 1000)
-        FString AmountText;
-        if (Amount >= 1000)
-        {
-            // Liquid: show m³ and per-minute in m³
-            int32 DisplayAmount = Amount / 1000;
-            float PerMinuteM3 = PerMinute / 1000.0f;
-            float GridTotalM3 = GridTotalPerMinute / 1000.0f;
-            AmountText = FText::Format(LOCTEXT("Panel_Recipe_OutputLiquid", "\u2192 Output: {0} x{1} m\u00B3 ({2}/min) (Grid Total: {3}/min)"), FText::FromString(ItemName), FText::AsNumber(DisplayAmount), FText::AsNumber(PerMinuteM3), FText::AsNumber(GridTotalM3)).ToString();
-        }
-        else
-        {
-            AmountText = FText::Format(LOCTEXT("Panel_Recipe_Output", "\u2192 Output: {0} x{1} ({2}/min) (Grid Total: {3}/min)"), FText::FromString(ItemName), FText::AsNumber(Amount), FText::AsNumber(PerMinute), FText::AsNumber(GridTotalPerMinute)).ToString();
-        }
-
-        UWidget* OutputRow = CreateRecipeDetailRow(ItemIcon, AmountText, OutputColor);
-        if (OutputRow)
-        {
-            RecipeDetailsContainer->AddChild(OutputRow);
-        }
-    }
-
-    // Add input rows with icons
-    TArray<FItemAmount> Ingredients = RecipeCDO->GetIngredients();
-    for (int32 i = 0; i < Ingredients.Num(); i++)
-    {
-        TSubclassOf<UFGItemDescriptor> ItemClass = Ingredients[i].ItemClass;
-        int32 Amount = Ingredients[i].Amount;
-        FString ItemName = UFGItemDescriptor::GetItemName(ItemClass).ToString();
-
-        // Get item icon
-        UTexture2D* ItemIcon = UFGItemDescriptor::GetSmallIcon(ItemClass);
-
-        // Calculate per-minute rate (single building) and grid total
-        float PerMinute = Amount * CyclesPerMinute;
-        float GridTotalPerMinute = PerMinute * GridTotal;
-
-        // Format amount with per-minute rate and grid total (liquids have amounts >= 1000)
-        FString AmountText;
-        if (Amount >= 1000)
-        {
-            // Liquid: show m³ and per-minute in m³
-            int32 DisplayAmount = Amount / 1000;
-            float PerMinuteM3 = PerMinute / 1000.0f;
-            float GridTotalM3 = GridTotalPerMinute / 1000.0f;
-            AmountText = FText::Format(LOCTEXT("Panel_Recipe_InputLiquid", "\u2190 Input: {0} x{1} m\u00B3 ({2}/min) (Grid Total: {3}/min)"), FText::FromString(ItemName), FText::AsNumber(DisplayAmount), FText::AsNumber(PerMinuteM3), FText::AsNumber(GridTotalM3)).ToString();
-        }
-        else
-        {
-            AmountText = FText::Format(LOCTEXT("Panel_Recipe_Input", "\u2190 Input: {0} x{1} ({2}/min) (Grid Total: {3}/min)"), FText::FromString(ItemName), FText::AsNumber(Amount), FText::AsNumber(PerMinute), FText::AsNumber(GridTotalPerMinute)).ToString();
-        }
-
-        UWidget* InputRow = CreateRecipeDetailRow(ItemIcon, AmountText, InputColor);
-        if (InputRow)
-        {
-            RecipeDetailsContainer->AddChild(InputRow);
-        }
-    }
-}
-
-UWidget* USmartSettingsFormWidget::CreateRecipeDetailRow(UTexture2D* Icon, const FString& Text, const FLinearColor& TextColor)
-{
-    // Create horizontal box for the row
-    UHorizontalBox* Row = NewObject<UHorizontalBox>(this);
-    if (!Row)
-    {
-        return nullptr;
-    }
-
-    // Add icon if provided
-    if (Icon)
-    {
-        USizeBox* IconSizeBox = NewObject<USizeBox>(this);
-        if (IconSizeBox)
-        {
-            IconSizeBox->SetWidthOverride(24.0f);
-            IconSizeBox->SetHeightOverride(24.0f);
-
-            UImage* IconImage = NewObject<UImage>(this);
-            if (IconImage)
-            {
-                IconImage->SetBrushFromTexture(Icon);
-                IconSizeBox->AddChild(IconImage);
-            }
-
-            UHorizontalBoxSlot* IconSlot = Row->AddChildToHorizontalBox(IconSizeBox);
-            if (IconSlot)
-            {
-                IconSlot->SetPadding(FMargin(0.0f, 0.0f, 8.0f, 0.0f));  // Right padding
-                IconSlot->SetVerticalAlignment(VAlign_Center);
-            }
-        }
-    }
-
-    // Add text
-    UTextBlock* TextWidget = NewObject<UTextBlock>(this);
-    if (TextWidget)
-    {
-        TextWidget->SetText(FText::FromString(Text));
-        TextWidget->SetColorAndOpacity(FSlateColor(TextColor));
-
-        // Set font to match HUD style (in-game multi-script font)
-        TextWidget->SetFont(SFFont::Get(14));
-
-        UHorizontalBoxSlot* TextSlot = Row->AddChildToHorizontalBox(TextWidget);
-        if (TextSlot)
-        {
-            TextSlot->SetVerticalAlignment(VAlign_Center);
-        }
-    }
-
-    return Row;
-}
-
-// ============================================================================
-// Belt Auto-Connect Controls
-// ============================================================================
-
-FString USmartSettingsFormWidget::BeltTierToDisplayString(int32 Tier) const
-{
-    if (Tier <= 0)
-    {
-        // Show what "Auto" resolves to
-        if (CachedSubsystem.IsValid())
-        {
-            int32 HighestTier = CachedSubsystem->GetHighestUnlockedBeltTier(
-                Cast<AFGPlayerController>(GetOwningPlayer()));
-            return FString::Printf(TEXT("Auto (Mk%d)"), HighestTier);
-        }
-        return TEXT("Auto");
-    }
-    return FString::Printf(TEXT("Mk%d"), Tier);
-}
-
-int32 USmartSettingsFormWidget::DisplayStringToBeltTier(const FString& DisplayString) const
-{
-    if (DisplayString.StartsWith(TEXT("Auto")))
-    {
-        return 0;
-    }
-
-    // Parse "Mk1", "Mk2", etc.
-    if (DisplayString.StartsWith(TEXT("Mk")))
-    {
-        FString TierStr = DisplayString.RightChop(2);
-        return FCString::Atoi(*TierStr);
-    }
-
-    return 0;
-}
-
-void USmartSettingsFormWidget::PopulateBeltTierComboBoxes()
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    // Get highest unlocked belt tier
-    AFGPlayerController* PC = Cast<AFGPlayerController>(GetOwningPlayer());
-    int32 HighestUnlocked = CachedSubsystem->GetHighestUnlockedBeltTier(PC);
-
-    // Build list of available tiers: Auto (0), then Mk1 through highest unlocked
-    CachedBeltTiers.Empty();
-    CachedBeltTiers.Add(0);  // Auto
-    for (int32 Tier = 1; Tier <= HighestUnlocked; Tier++)
-    {
-        CachedBeltTiers.Add(Tier);
-    }
-
-    // Populate Main tier ComboBox
-    if (BeltTierMainComboBox)
-    {
-        BeltTierMainComboBox->ClearOptions();
-        for (int32 Tier : CachedBeltTiers)
-        {
-            BeltTierMainComboBox->AddOption(BeltTierToDisplayString(Tier));
-        }
-    }
-
-    // Populate To Building tier ComboBox
-    if (BeltTierToBuildingComboBox)
-    {
-        BeltTierToBuildingComboBox->ClearOptions();
-        for (int32 Tier : CachedBeltTiers)
-        {
-            BeltTierToBuildingComboBox->AddOption(BeltTierToDisplayString(Tier));
-        }
-    }
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Populated belt tier ComboBoxes with %d options (highest unlocked: Mk%d)"),
-        CachedBeltTiers.Num(), HighestUnlocked);
-}
-
-void USmartSettingsFormWidget::UpdateBeltAutoConnectControls()
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    const auto& Settings = CachedSubsystem->GetAutoConnectRuntimeSettings();
-
-    // Update Enabled checkbox
-    if (BeltEnabledCheckBox)
-    {
-        BeltEnabledCheckBox->SetIsChecked(Settings.bEnabled);
-    }
-
-    // Update Main tier ComboBox
-    if (BeltTierMainComboBox)
-    {
-        int32 Index = CachedBeltTiers.Find(Settings.BeltTierMain);
-        if (Index == INDEX_NONE)
-        {
-            Index = 0;  // Default to Auto if current setting not in list
-        }
-        BeltTierMainComboBox->SetSelectedIndex(Index);
-    }
-
-    // Update To Building tier ComboBox
-    if (BeltTierToBuildingComboBox)
-    {
-        // Restore original event binding (in case it was repurposed for stackable poles)
-        BeltTierToBuildingComboBox->OnSelectionChanged.Clear();
-        BeltTierToBuildingComboBox->OnSelectionChanged.AddDynamic(this, &USmartSettingsFormWidget::OnBeltTierToBuildingChanged);
-
-        int32 Index = CachedBeltTiers.Find(Settings.BeltTierToBuilding);
-        if (Index == INDEX_NONE)
-        {
-            Index = 0;  // Default to Auto if current setting not in list
-        }
-        BeltTierToBuildingComboBox->SetSelectedIndex(Index);
-    }
-
-    // Update Chain checkbox
-    if (BeltChainCheckBox)
-    {
-        BeltChainCheckBox->SetIsChecked(Settings.bChainDistributors);
-        BeltChainCheckBox->SetVisibility(ESlateVisibility::Visible);
-    }
-
-    // Restore visibility for regular belt controls, hide stackable-only controls
-    if (BeltTierToBuildingLabel)
-    {
-        BeltTierToBuildingLabel->SetVisibility(ESlateVisibility::Visible);
-    }
-    if (BeltTierToBuildingComboBox)
-    {
-        BeltTierToBuildingComboBox->SetVisibility(ESlateVisibility::Visible);
-    }
-    if (StackableBeltDirectionRow)
-    {
-        StackableBeltDirectionRow->SetVisibility(ESlateVisibility::Collapsed);
-    }
-    if (BeltChainLabel)
-    {
-        BeltChainLabel->SetVisibility(ESlateVisibility::Visible);
-    }
-
-    // Update Routing Mode ComboBox
-    if (BeltRoutingModeComboBox)
-    {
-        BeltRoutingModeComboBox->ClearOptions();
-        BeltRoutingModeComboBox->AddOption(TEXT("Default"));
-        BeltRoutingModeComboBox->AddOption(TEXT("Curve"));
-        BeltRoutingModeComboBox->AddOption(TEXT("Straight"));
-
-        int32 ModeIndex = FMath::Clamp(Settings.BeltRoutingMode, 0, 2);
-        BeltRoutingModeComboBox->SetSelectedIndex(ModeIndex);
-    }
-    if (BeltRoutingModeRow)
-    {
-        BeltRoutingModeRow->SetVisibility(ESlateVisibility::Visible);
-    }
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Updated belt controls - Enabled=%d, Main=%d, ToBuilding=%d, Chain=%d, RoutingMode=%d"),
-        Settings.bEnabled, Settings.BeltTierMain, Settings.BeltTierToBuilding, Settings.bChainDistributors, Settings.BeltRoutingMode);
-}
-
-void USmartSettingsFormWidget::OnBeltEnabledChanged(bool bIsChecked)
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    CachedSubsystem->SetAutoConnectBeltEnabled(bIsChecked);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Belt auto-connect enabled changed to %d"), bIsChecked);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-void USmartSettingsFormWidget::OnBeltTierMainChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    if (SelectionType == ESelectInfo::Direct)
-    {
-        return;  // Ignore programmatic changes
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    int32 Tier = DisplayStringToBeltTier(SelectedItem);
-    CachedSubsystem->SetAutoConnectBeltTierMain(Tier);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Belt tier main changed to %d (%s)"), Tier, *SelectedItem);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-void USmartSettingsFormWidget::OnBeltTierToBuildingChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    if (SelectionType == ESelectInfo::Direct)
-    {
-        return;  // Ignore programmatic changes
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    int32 Tier = DisplayStringToBeltTier(SelectedItem);
-    CachedSubsystem->SetAutoConnectBeltTierToBuilding(Tier);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Belt tier to building changed to %d (%s)"), Tier, *SelectedItem);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-void USmartSettingsFormWidget::OnBeltChainChanged(bool bIsChecked)
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    CachedSubsystem->SetAutoConnectBeltChain(bIsChecked);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Belt chain changed to %d"), bIsChecked);
-}
-
-void USmartSettingsFormWidget::OnStackableBeltDirectionChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    if (SelectionType == ESelectInfo::Direct)
-    {
-        return;  // Ignore programmatic changes
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    // Convert "Forward"/"Backward" to 0/1
-    int32 Direction = 0;  // Default to Forward
-    if (SelectedItem == TEXT("Backward"))
-    {
-        Direction = 1;
-    }
-
-    CachedSubsystem->SetAutoConnectStackableBeltDirection(Direction);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Stackable belt direction changed to %d (%s)"), Direction, *SelectedItem);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-void USmartSettingsFormWidget::OnBeltRoutingModeChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    if (SelectionType == ESelectInfo::Direct)
-    {
-        return;  // Ignore programmatic changes
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    // Convert display string to routing mode index
-    int32 RoutingMode = 0;  // Default
-    if (SelectedItem == TEXT("Curve")) RoutingMode = 1;
-    else if (SelectedItem == TEXT("Straight")) RoutingMode = 2;
-
-    CachedSubsystem->SetAutoConnectBeltRoutingMode(RoutingMode);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Belt routing mode changed to %d (%s)"), RoutingMode, *SelectedItem);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-// ============================================================================
-// Pipe Auto-Connect Controls
-// ============================================================================
-
-FString USmartSettingsFormWidget::PipeTierToDisplayString(int32 Tier) const
-{
-    if (Tier <= 0)
-    {
-        // Show what "Auto" resolves to
-        if (CachedSubsystem.IsValid())
-        {
-            int32 HighestTier = CachedSubsystem->GetHighestUnlockedPipeTier(
-                Cast<AFGPlayerController>(GetOwningPlayer()));
-            return FString::Printf(TEXT("Auto (Mk%d)"), HighestTier);
-        }
-        return TEXT("Auto");
-    }
-    return FString::Printf(TEXT("Mk%d"), Tier);
-}
-
-int32 USmartSettingsFormWidget::DisplayStringToPipeTier(const FString& DisplayString) const
-{
-    if (DisplayString.StartsWith(TEXT("Auto")))
-    {
-        return 0;
-    }
-
-    // Parse "Mk1", "Mk2"
-    if (DisplayString.StartsWith(TEXT("Mk")))
-    {
-        FString TierStr = DisplayString.RightChop(2);
-        return FCString::Atoi(*TierStr);
-    }
-
-    return 0;
-}
-
-void USmartSettingsFormWidget::PopulatePipeTierComboBoxes()
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    // Get highest unlocked pipe tier
-    AFGPlayerController* PC = Cast<AFGPlayerController>(GetOwningPlayer());
-    int32 HighestUnlocked = CachedSubsystem->GetHighestUnlockedPipeTier(PC);
-
-    // Build list of available tiers: Auto (0), then Mk1 through highest unlocked
-    CachedPipeTiers.Empty();
-    CachedPipeTiers.Add(0);  // Auto
-    for (int32 Tier = 1; Tier <= HighestUnlocked; Tier++)
-    {
-        CachedPipeTiers.Add(Tier);
-    }
-
-    // Populate Main tier ComboBox
-    if (PipeTierMainComboBox)
-    {
-        PipeTierMainComboBox->ClearOptions();
-        for (int32 Tier : CachedPipeTiers)
-        {
-            PipeTierMainComboBox->AddOption(PipeTierToDisplayString(Tier));
-        }
-    }
-
-    // Populate To Building tier ComboBox
-    if (PipeTierToBuildingComboBox)
-    {
-        PipeTierToBuildingComboBox->ClearOptions();
-        for (int32 Tier : CachedPipeTiers)
-        {
-            PipeTierToBuildingComboBox->AddOption(PipeTierToDisplayString(Tier));
-        }
-    }
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Populated pipe tier ComboBoxes with %d options (highest unlocked: Mk%d)"),
-        CachedPipeTiers.Num(), HighestUnlocked);
-}
-
-void USmartSettingsFormWidget::UpdatePipeAutoConnectControls()
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    const auto& Settings = CachedSubsystem->GetAutoConnectRuntimeSettings();
-
-    // Update Enabled checkbox
-    if (PipeEnabledCheckBox)
-    {
-        PipeEnabledCheckBox->SetIsChecked(Settings.bPipeAutoConnectEnabled);
-    }
-
-    // Update Main tier ComboBox
-    if (PipeTierMainComboBox)
-    {
-        int32 Index = CachedPipeTiers.Find(Settings.PipeTierMain);
-        if (Index == INDEX_NONE)
-        {
-            Index = 0;  // Default to Auto if current setting not in list
-        }
-        PipeTierMainComboBox->SetSelectedIndex(Index);
-    }
-
-    // Update To Building tier ComboBox
-    if (PipeTierToBuildingComboBox)
-    {
-        int32 Index = CachedPipeTiers.Find(Settings.PipeTierToBuilding);
-        if (Index == INDEX_NONE)
-        {
-            Index = 0;  // Default to Auto if current setting not in list
-        }
-        PipeTierToBuildingComboBox->SetSelectedIndex(Index);
-    }
-
-    // Ensure both rows are visible for standard pipe junctions
-    if (PipeTierMainRow) PipeTierMainRow->SetVisibility(ESlateVisibility::Visible);
-    if (PipeTierToBuildingRow) PipeTierToBuildingRow->SetVisibility(ESlateVisibility::Visible);
-
-    // Update Flow Indicator checkbox and row visibility
-    // Only show if clean pipes are unlocked - otherwise flow indicators are implicit
-    AFGPlayerController* PC = Cast<AFGPlayerController>(GetOwningPlayer());
-    bool bCleanPipesUnlocked = CachedSubsystem->AreCleanPipesUnlocked(PC);
-
-    if (PipeIndicatorRow)
-    {
-        PipeIndicatorRow->SetVisibility(bCleanPipesUnlocked ? ESlateVisibility::Visible : ESlateVisibility::Collapsed);
-    }
-
-    if (PipeIndicatorCheckBox)
-    {
-        // If clean pipes not unlocked, force to true (with indicators)
-        if (!bCleanPipesUnlocked)
-        {
-            PipeIndicatorCheckBox->SetIsChecked(true);
-        }
-        else
-        {
-            PipeIndicatorCheckBox->SetIsChecked(Settings.bPipeIndicator);
-        }
-    }
-
-    // Update Routing Mode ComboBox
-    if (PipeRoutingModeComboBox)
-    {
-        PipeRoutingModeComboBox->ClearOptions();
-        PipeRoutingModeComboBox->AddOption(TEXT("Auto"));
-        PipeRoutingModeComboBox->AddOption(TEXT("Auto 2D"));
-        PipeRoutingModeComboBox->AddOption(TEXT("Straight"));
-        PipeRoutingModeComboBox->AddOption(TEXT("Curve"));
-        PipeRoutingModeComboBox->AddOption(TEXT("Noodle"));
-        PipeRoutingModeComboBox->AddOption(TEXT("Horiz→Vert"));
-
-        int32 ModeIndex = FMath::Clamp(Settings.PipeRoutingMode, 0, 5);
-        PipeRoutingModeComboBox->SetSelectedIndex(ModeIndex);
-    }
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Updated pipe controls - Enabled=%d, Main=%d, ToBuilding=%d, Indicator=%d, RoutingMode=%d, CleanUnlocked=%d"),
-        Settings.bPipeAutoConnectEnabled, Settings.PipeTierMain, Settings.PipeTierToBuilding, Settings.bPipeIndicator, Settings.PipeRoutingMode, bCleanPipesUnlocked);
-}
-
-void USmartSettingsFormWidget::OnPipeEnabledChanged(bool bIsChecked)
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    CachedSubsystem->SetAutoConnectPipeEnabled(bIsChecked);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Pipe auto-connect enabled changed to %d"), bIsChecked);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-void USmartSettingsFormWidget::OnPipeTierMainChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    if (SelectionType == ESelectInfo::Direct)
-    {
-        return;  // Ignore programmatic changes
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    int32 Tier = DisplayStringToPipeTier(SelectedItem);
-    CachedSubsystem->SetAutoConnectPipeTierMain(Tier);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Pipe tier main changed to %d (%s)"), Tier, *SelectedItem);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-void USmartSettingsFormWidget::OnPipeTierToBuildingChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    if (SelectionType == ESelectInfo::Direct)
-    {
-        return;  // Ignore programmatic changes
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    int32 Tier = DisplayStringToPipeTier(SelectedItem);
-    CachedSubsystem->SetAutoConnectPipeTierToBuilding(Tier);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Pipe tier to building changed to %d (%s)"), Tier, *SelectedItem);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-void USmartSettingsFormWidget::OnPipeIndicatorChanged(bool bIsChecked)
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    CachedSubsystem->SetAutoConnectPipeIndicator(bIsChecked);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Pipe indicator changed to %d (%s)"), bIsChecked, bIsChecked ? TEXT("Normal") : TEXT("Clean"));
-}
-
-void USmartSettingsFormWidget::OnPipeRoutingModeChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    if (SelectionType == ESelectInfo::Direct)
-    {
-        return;  // Ignore programmatic changes
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    // Convert display string to routing mode index
-    int32 RoutingMode = 0;
-    if (SelectedItem == TEXT("Auto")) RoutingMode = 0;
-    else if (SelectedItem == TEXT("Auto 2D")) RoutingMode = 1;
-    else if (SelectedItem == TEXT("Straight")) RoutingMode = 2;
-    else if (SelectedItem == TEXT("Curve")) RoutingMode = 3;
-    else if (SelectedItem == TEXT("Noodle")) RoutingMode = 4;
-    else if (SelectedItem == TEXT("Horiz→Vert")) RoutingMode = 5;
-
-    CachedSubsystem->SetAutoConnectPipeRoutingMode(RoutingMode);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Pipe routing mode changed to %d (%s)"), RoutingMode, *SelectedItem);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-// ============================================================================
-// Power Auto-Connect Controls
-// ============================================================================
-
-FString USmartSettingsFormWidget::PowerGridAxisToDisplayString(int32 Axis) const
-{
-    switch (Axis)
-    {
-    case 0: return TEXT("Auto");
-    case 1: return TEXT("X");
-    case 2: return TEXT("Y");
-    case 3: return TEXT("X+Y");
-    default: return TEXT("Auto");
-    }
-}
-
-int32 USmartSettingsFormWidget::DisplayStringToPowerGridAxis(const FString& DisplayString) const
-{
-    if (DisplayString == TEXT("Auto")) return 0;
-    if (DisplayString == TEXT("X")) return 1;
-    if (DisplayString == TEXT("Y")) return 2;
-    if (DisplayString == TEXT("X+Y")) return 3;
-    return 0;
-}
-
-void USmartSettingsFormWidget::PopulatePowerComboBoxes()
-{
-    // Populate Grid Axis ComboBox
-    if (PowerGridAxisComboBox)
-    {
-        PowerGridAxisComboBox->ClearOptions();
-        PowerGridAxisComboBox->AddOption(TEXT("Auto"));
-        PowerGridAxisComboBox->AddOption(TEXT("X"));
-        PowerGridAxisComboBox->AddOption(TEXT("Y"));
-        PowerGridAxisComboBox->AddOption(TEXT("X+Y"));
-    }
-
-    // Populate Reserved Slots ComboBox (0-5)
-    if (PowerReservedComboBox)
-    {
-        PowerReservedComboBox->ClearOptions();
-        for (int32 i = 0; i <= 5; i++)
-        {
-            PowerReservedComboBox->AddOption(FString::Printf(TEXT("%d"), i));
-        }
-    }
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Populated power ComboBoxes"));
-}
-
-void USmartSettingsFormWidget::UpdatePowerAutoConnectControls()
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    const auto& Settings = CachedSubsystem->GetAutoConnectRuntimeSettings();
-
-    // Update Enabled checkbox
-    if (PowerEnabledCheckBox)
-    {
-        PowerEnabledCheckBox->SetIsChecked(Settings.bConnectPower);
-    }
-
-    // Update Grid Axis ComboBox
-    if (PowerGridAxisComboBox)
-    {
-        PowerGridAxisComboBox->SetSelectedOption(PowerGridAxisToDisplayString(Settings.PowerGridAxis));
-    }
-
-    // Update Reserved Slots ComboBox
-    if (PowerReservedComboBox)
-    {
-        PowerReservedComboBox->SetSelectedOption(FString::Printf(TEXT("%d"), Settings.PowerReserved));
-    }
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Updated power controls - Enabled=%d, GridAxis=%d, Reserved=%d"),
-        Settings.bConnectPower, Settings.PowerGridAxis, Settings.PowerReserved);
-}
-
-void USmartSettingsFormWidget::OnPowerEnabledChanged(bool bIsChecked)
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    CachedSubsystem->SetAutoConnectPowerEnabled(bIsChecked);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Power auto-connect enabled changed to %d"), bIsChecked);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-void USmartSettingsFormWidget::OnPowerGridAxisChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    if (SelectionType == ESelectInfo::Direct)
-    {
-        return;  // Ignore programmatic changes
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    int32 Axis = DisplayStringToPowerGridAxis(SelectedItem);
-    CachedSubsystem->SetAutoConnectPowerGridAxis(Axis);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Power grid axis changed to %d (%s)"), Axis, *SelectedItem);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-void USmartSettingsFormWidget::OnPowerReservedChanged(FString SelectedItem, ESelectInfo::Type SelectionType)
-{
-    if (SelectionType == ESelectInfo::Direct)
-    {
-        return;  // Ignore programmatic changes
-    }
-
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    int32 Reserved = FCString::Atoi(*SelectedItem);
-    CachedSubsystem->SetAutoConnectPowerReserved(Reserved);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Power reserved slots changed to %d"), Reserved);
-
-    if (bApplyImmediately)
-    {
-        CachedSubsystem->TriggerAutoConnectRefresh();
-    }
-}
-
-// ========================================
-// Grid Direction Toggle Handlers
-// ========================================
-
-void USmartSettingsFormWidget::OnGridXDirToggleClicked()
-{
-    bGridXPositive = !bGridXPositive;
-    UpdateGridDirectionLabel(GridXDirLabel, bGridXPositive);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Grid X direction toggled to %s"), bGridXPositive ? TEXT("+") : TEXT("-"));
-}
-
-void USmartSettingsFormWidget::OnGridYDirToggleClicked()
-{
-    bGridYPositive = !bGridYPositive;
-    UpdateGridDirectionLabel(GridYDirLabel, bGridYPositive);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Grid Y direction toggled to %s"), bGridYPositive ? TEXT("+") : TEXT("-"));
-}
-
-void USmartSettingsFormWidget::OnGridZDirToggleClicked()
-{
-    bGridZPositive = !bGridZPositive;
-    UpdateGridDirectionLabel(GridZDirLabel, bGridZPositive);
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Grid Z direction toggled to %s"), bGridZPositive ? TEXT("+") : TEXT("-"));
-}
-
-void USmartSettingsFormWidget::UpdateGridDirectionLabel(UTextBlock* Label, bool bPositive)
-{
-    if (Label)
-    {
-        Label->SetText(FText::FromString(bPositive ? TEXT("+") : TEXT("-")));
-    }
-}
-
-// ========================================
-// Keyboard Input Handling
-// ========================================
-
-void USmartSettingsFormWidget::NativeDestruct()
-{
-    Super::NativeDestruct();
-}
-
-FReply USmartSettingsFormWidget::NativeOnKeyDown(const FGeometry& InGeometry, const FKeyEvent& InKeyEvent)
-{
-    // Escape key behavior depends on state
-    if (InKeyEvent.GetKey() == EKeys::Escape)
-    {
-        // If confirmation dialog is open, dismiss it
-        if (bWaitingForConfirmation)
-        {
-            UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Escape pressed - dismissing confirmation dialog"));
-            OnConfirmNoClicked();
-            return FReply::Handled();
-        }
-
-        // Otherwise, cancel unapplied changes and close form
-        UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Escape pressed - reverting and closing"));
-        RevertToLastAppliedState();
-        CloseForm();
-        return FReply::Handled();
-    }
-
-    // Enter key applies changes (only when a SpinBox has focus)
-    if (InKeyEvent.GetKey() == EKeys::Enter && IsAnyTextInputFocused())
-    {
-        UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Enter pressed with SpinBox focused - applying"));
-        OnApplyButtonClicked();
-        return FReply::Handled();
-    }
-
-    // Issue #212: Tab/Shift+Tab navigates between SpinBox fields in logical order
-    if (InKeyEvent.GetKey() == EKeys::Tab && IsAnyTextInputFocused())
-    {
-        // Build ordered list of all SpinBoxes for tab navigation
-        TArray<USpinBox*> TabOrder;
-        if (GridXInput) TabOrder.Add(GridXInput);
-        if (GridYInput) TabOrder.Add(GridYInput);
-        if (GridZInput) TabOrder.Add(GridZInput);
-        if (SpacingXInput) TabOrder.Add(SpacingXInput);
-        if (SpacingYInput) TabOrder.Add(SpacingYInput);
-        if (SpacingZInput) TabOrder.Add(SpacingZInput);
-        if (StepsXInput) TabOrder.Add(StepsXInput);
-        if (StepsYInput) TabOrder.Add(StepsYInput);
-        if (StaggerXInput) TabOrder.Add(StaggerXInput);
-        if (StaggerYInput) TabOrder.Add(StaggerYInput);
-        if (StaggerZXInput) TabOrder.Add(StaggerZXInput);
-        if (StaggerZYInput) TabOrder.Add(StaggerZYInput);
-        if (RotationZInput) TabOrder.Add(RotationZInput);
-
-        // Find currently focused SpinBox
-        int32 CurrentIndex = INDEX_NONE;
-        for (int32 i = 0; i < TabOrder.Num(); ++i)
-        {
-            if (TabOrder[i]->HasKeyboardFocus())
-            {
-                CurrentIndex = i;
-                break;
-            }
-        }
-
-        if (CurrentIndex != INDEX_NONE && TabOrder.Num() > 1)
-        {
-            int32 NextIndex;
-            if (InKeyEvent.IsShiftDown())
-            {
-                // Shift+Tab: go backward (wrap around)
-                NextIndex = (CurrentIndex - 1 + TabOrder.Num()) % TabOrder.Num();
-            }
-            else
-            {
-                // Tab: go forward (wrap around)
-                NextIndex = (CurrentIndex + 1) % TabOrder.Num();
-            }
-
-            TabOrder[NextIndex]->SetKeyboardFocus();
-            UE_LOG(LogSmartFoundations, Verbose, TEXT("Settings Form: Tab navigation %d -> %d"), CurrentIndex, NextIndex);
-            return FReply::Handled();
-        }
-    }
-
-    return Super::NativeOnKeyDown(InGeometry, InKeyEvent);
-}
-
-// ========================================
-// SpinBox Value Handlers
-// ========================================
-
-void USmartSettingsFormWidget::OnSpinBoxValueChanged(float Value)
-{
-    // Called during slider drag or mouse wheel - only apply in immediate mode
-    if (bApplyImmediately)
-    {
-        UE_LOG(LogSmartFoundations, Verbose, TEXT("Settings Form: SpinBox value changed (immediate mode)"));
-        OnApplyButtonClicked();
-    }
-}
-
-void USmartSettingsFormWidget::OnGridSpinBoxValueChanged(float Value)
-{
-    // Update warning display whenever grid values change
-    UpdateGridWarningDisplay();
-
-    // Also handle immediate mode apply (if still enabled - will be auto-disabled for large grids)
-    if (bApplyImmediately)
-    {
-        UE_LOG(LogSmartFoundations, Verbose, TEXT("Settings Form: Grid SpinBox value changed (immediate mode)"));
-        OnApplyButtonClicked();
-    }
-}
-
-void USmartSettingsFormWidget::OnSpinBoxValueCommitted(float Value, ETextCommit::Type CommitMethod)
-{
-    // Enter key is now handled globally in NativeOnKeyDown to avoid false triggers
-    // from SpinBox button clicks which also fire OnEnter
-
-    // Only apply on focus loss in immediate mode
-    if (bApplyImmediately && (CommitMethod == ETextCommit::Default || CommitMethod == ETextCommit::OnUserMovedFocus))
-    {
-        UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: SpinBox focus lost (immediate mode) - applying"));
-        OnApplyButtonClicked();
-    }
-}
-
-// ========================================
-// Apply Immediately Checkbox Handler
-// ========================================
-
-void USmartSettingsFormWidget::OnApplyImmediatelyChanged(bool bIsChecked)
-{
-    bApplyImmediately = bIsChecked;
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Apply immediately mode changed to %d"), bIsChecked);
-}
-
-// ========================================
-// State Cache/Revert Helpers
-// ========================================
-
-void USmartSettingsFormWidget::CacheCurrentStateAsApplied()
-{
-    if (CachedSubsystem.IsValid())
-    {
-        LastAppliedState = CachedSubsystem->GetCounterState();
-        bLastAppliedGridXPositive = bGridXPositive;
-        bLastAppliedGridYPositive = bGridYPositive;
-        bLastAppliedGridZPositive = bGridZPositive;
-    }
-}
-
-void USmartSettingsFormWidget::RevertToLastAppliedState()
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    // Restore the last applied state
-    CachedSubsystem->UpdateCounterState(LastAppliedState);
-    CachedSubsystem->RegenerateChildHologramGrid();
-
-    // Restore direction state
-    bGridXPositive = bLastAppliedGridXPositive;
-    bGridYPositive = bLastAppliedGridYPositive;
-    bGridZPositive = bLastAppliedGridZPositive;
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Reverted to last applied state"));
-}
-
-bool USmartSettingsFormWidget::IsAnyTextInputFocused() const
-{
-    auto IsSpinBoxFocused = [](USpinBox* SpinBox) -> bool
-    {
-        return SpinBox && SpinBox->HasKeyboardFocus();
-    };
-
-    return IsSpinBoxFocused(GridXInput) ||
-           IsSpinBoxFocused(GridYInput) ||
-           IsSpinBoxFocused(GridZInput) ||
-           IsSpinBoxFocused(SpacingXInput) ||
-           IsSpinBoxFocused(SpacingYInput) ||
-           IsSpinBoxFocused(SpacingZInput) ||
-           IsSpinBoxFocused(StepsXInput) ||
-           IsSpinBoxFocused(StepsYInput) ||
-           IsSpinBoxFocused(StaggerXInput) ||
-           IsSpinBoxFocused(StaggerYInput) ||
-           IsSpinBoxFocused(StaggerZXInput) ||
-           IsSpinBoxFocused(StaggerZYInput) ||
-           IsSpinBoxFocused(RotationZInput);
-}
-
-// ========================================
-// Large Grid Warning System
-// ========================================
-
-int32 USmartSettingsFormWidget::CalculateGridTotal() const
-{
-    int32 GridX = GridXInput ? FMath::Max(1, FMath::RoundToInt(GridXInput->GetValue())) : 1;
-    int32 GridY = GridYInput ? FMath::Max(1, FMath::RoundToInt(GridYInput->GetValue())) : 1;
-    int32 GridZ = GridZInput ? FMath::Max(1, FMath::RoundToInt(GridZInput->GetValue())) : 1;
-
-    return GridX * GridY * GridZ;
-}
-
-void USmartSettingsFormWidget::UpdateGridWarningDisplay()
-{
-    int32 GridTotal = CalculateGridTotal();
-
-    // Update Grid Total text
-    if (GridTotalText)
-    {
-        GridTotalText->SetText(FText::Format(LOCTEXT("Panel_GridTotal", "Grid Total: {0} holograms"), FText::AsNumber(GridTotal)));
-
-        // Set color based on warning level
-        FLinearColor TextColor;
-        if (GridTotal >= GRID_WARNING_THRESHOLD_DANGER)
-        {
-            TextColor = FLinearColor(1.0f, 0.2f, 0.2f, 1.0f);  // Red
-        }
-        else if (GridTotal >= GRID_WARNING_THRESHOLD_WARNING)
-        {
-            TextColor = FLinearColor(1.0f, 0.5f, 0.0f, 1.0f);  // Orange
-        }
-        else if (GridTotal >= GRID_WARNING_THRESHOLD_CAUTION)
-        {
-            TextColor = FLinearColor(1.0f, 0.85f, 0.0f, 1.0f); // Yellow
-        }
-        else
-        {
-            TextColor = FLinearColor(1.0f, 1.0f, 1.0f, 1.0f);  // White
-        }
-        GridTotalText->SetColorAndOpacity(FSlateColor(TextColor));
-    }
-
-    // Update warning text (avoid Unicode symbols - game font doesn't support them)
-    if (GridWarningText)
-    {
-        if (GridTotal >= GRID_WARNING_THRESHOLD_DANGER)
-        {
-            GridWarningText->SetText(LOCTEXT("Panel_GridWarning_Danger", "* Extreme grid - high risk of freeze/crash"));
-            GridWarningText->SetColorAndOpacity(FSlateColor(FLinearColor(1.0f, 0.2f, 0.2f, 1.0f)));
-            GridWarningText->SetVisibility(ESlateVisibility::Visible);
-        }
-        else if (GridTotal >= GRID_WARNING_THRESHOLD_WARNING)
-        {
-            GridWarningText->SetText(LOCTEXT("Panel_GridWarning_Warning", "* Very large grid - may cause temporary freeze"));
-            GridWarningText->SetColorAndOpacity(FSlateColor(FLinearColor(1.0f, 0.5f, 0.0f, 1.0f)));
-            GridWarningText->SetVisibility(ESlateVisibility::Visible);
-        }
-        else if (GridTotal >= GRID_WARNING_THRESHOLD_CAUTION)
-        {
-            GridWarningText->SetText(LOCTEXT("Panel_GridWarning_Caution", "* Large grid"));
-            GridWarningText->SetColorAndOpacity(FSlateColor(FLinearColor(1.0f, 0.85f, 0.0f, 1.0f)));
-            GridWarningText->SetVisibility(ESlateVisibility::Visible);
-        }
-        else
-        {
-            GridWarningText->SetVisibility(ESlateVisibility::Collapsed);
-        }
-    }
-
-    // Update Apply Immediately state based on grid size
-    UpdateApplyImmediatelyState();
-}
-
-bool USmartSettingsFormWidget::RequiresApplyConfirmation() const
-{
-    return CalculateGridTotal() >= GRID_WARNING_THRESHOLD_WARNING;
-}
-
-void USmartSettingsFormWidget::UpdateApplyImmediatelyState()
-{
-    int32 GridTotal = CalculateGridTotal();
-
-    if (GridTotal >= GRID_WARNING_THRESHOLD_WARNING)
-    {
-        // Auto-disable Apply Immediately for large grids
-        if (bApplyImmediately)
-        {
-            bApplyImmediatelyWasEnabled = true;
-            bApplyImmediately = false;
-
-            if (ApplyImmediatelyCheckBox)
-            {
-                ApplyImmediatelyCheckBox->SetIsChecked(false);
-                ApplyImmediatelyCheckBox->SetIsEnabled(false);
-            }
-
-            UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Apply Immediately auto-disabled for large grid (%d objects)"), GridTotal);
-        }
-        else if (ApplyImmediatelyCheckBox)
-        {
-            // Keep it disabled even if user didn't have it enabled
-            ApplyImmediatelyCheckBox->SetIsEnabled(false);
-        }
-    }
-    else
-    {
-        // Re-enable Apply Immediately checkbox
-        if (ApplyImmediatelyCheckBox)
-        {
-            ApplyImmediatelyCheckBox->SetIsEnabled(true);
-        }
-
-        // Restore previous state if it was auto-disabled
-        if (bApplyImmediatelyWasEnabled)
-        {
-            bApplyImmediately = true;
-            bApplyImmediatelyWasEnabled = false;
-
-            if (ApplyImmediatelyCheckBox)
-            {
-                ApplyImmediatelyCheckBox->SetIsChecked(true);
-            }
-
-            UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Apply Immediately restored (grid now %d objects)"), GridTotal);
-        }
-    }
-}
-
-void USmartSettingsFormWidget::ShowLargeGridConfirmation(TFunction<void()> OnConfirmed)
-{
-    int32 GridTotal = CalculateGridTotal();
-
-    // Store callback for when user confirms
-    PendingConfirmCallback = OnConfirmed;
-    bWaitingForConfirmation = true;
-
-    // Determine warning level and message
-    FString Title;
-    FString Message;
-    FLinearColor TitleColor;
-
-    if (GridTotal >= GRID_WARNING_THRESHOLD_DANGER)
-    {
-        Title = LOCTEXT("Panel_Confirm_ExtremeTitle", "EXTREME GRID WARNING!").ToString();
-        Message = FText::Format(LOCTEXT("Panel_Confirm_ExtremeMsg", "You are about to generate {0} holograms.\n\nThis will likely cause a significant freeze and may crash the game.\n\nAre you sure you want to continue?"), FText::AsNumber(GridTotal)).ToString();
-        TitleColor = FLinearColor(1.0f, 0.2f, 0.2f, 1.0f);  // Red
-        UE_LOG(LogSmartFoundations, Warning, TEXT("Settings Form: Showing EXTREME grid confirmation for %d objects"), GridTotal);
-    }
-    else if (GridTotal >= GRID_WARNING_THRESHOLD_WARNING)
-    {
-        Title = LOCTEXT("Panel_Confirm_LargeTitle", "Large Grid Warning").ToString();
-        Message = FText::Format(LOCTEXT("Panel_Confirm_LargeMsg", "You are about to generate {0} holograms.\n\nThis may cause a temporary freeze.\n\nContinue?"), FText::AsNumber(GridTotal)).ToString();
-        TitleColor = FLinearColor(1.0f, 0.5f, 0.0f, 1.0f);  // Orange
-        UE_LOG(LogSmartFoundations, Warning, TEXT("Settings Form: Showing large grid confirmation for %d objects"), GridTotal);
-    }
-    else
-    {
-        // No confirmation needed, just proceed
-        bWaitingForConfirmation = false;
-        if (OnConfirmed)
-        {
-            OnConfirmed();
-        }
-        return;
-    }
-
-    // Show the confirmation dialog
-    ShowConfirmationDialog(Title, Message, TitleColor);
-}
-
-void USmartSettingsFormWidget::ShowConfirmationDialog(const FString& Title, const FString& Message, const FLinearColor& TitleColor)
-{
-    if (!ConfirmationSizeBox)
-    {
-        // No dialog widget available, just proceed
-        UE_LOG(LogSmartFoundations, Warning, TEXT("Settings Form: Confirmation dialog not available, proceeding anyway"));
-        if (PendingConfirmCallback)
-        {
-            PendingConfirmCallback();
-        }
-        bWaitingForConfirmation = false;
-        return;
-    }
-
-    // Set dialog content
-    if (ConfirmationTitle)
-    {
-        ConfirmationTitle->SetText(FText::FromString(Title));
-        ConfirmationTitle->SetColorAndOpacity(FSlateColor(TitleColor));
-    }
-
-    if (ConfirmationMessage)
-    {
-        ConfirmationMessage->SetText(FText::FromString(Message));
-    }
-
-    // Show the dialog (use SizeBox for visibility control)
-    ConfirmationSizeBox->SetVisibility(ESlateVisibility::Visible);
-
-    // Focus the Cancel button by default (safer option)
-    if (ConfirmNoButton)
-    {
-        ConfirmNoButton->SetKeyboardFocus();
-    }
-}
-
-void USmartSettingsFormWidget::HideConfirmationDialog()
-{
-    if (ConfirmationSizeBox)
-    {
-        ConfirmationSizeBox->SetVisibility(ESlateVisibility::Collapsed);
-    }
-    bWaitingForConfirmation = false;
-}
-
-void USmartSettingsFormWidget::OnConfirmYesClicked()
-{
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: User confirmed large grid operation"));
-
-    HideConfirmationDialog();
-
-    // Execute the pending callback
-    if (PendingConfirmCallback)
-    {
-        PendingConfirmCallback();
-        PendingConfirmCallback = nullptr;
-    }
-}
-
-void USmartSettingsFormWidget::OnConfirmNoClicked()
-{
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: User cancelled large grid operation"));
-
-    HideConfirmationDialog();
-    PendingConfirmCallback = nullptr;
-}
-
-// ============================================================================
-// Stackable Pole Auto-Connect Controls
-// ============================================================================
-
-void USmartSettingsFormWidget::UpdateStackableBeltAutoConnectControls()
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    const auto& Settings = CachedSubsystem->GetAutoConnectRuntimeSettings();
-
-    // Log widget availability for debugging
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Stackable Belt Widget Status: Direction=%s, TierToBuilding=%s, Chain=%s"),
-        StackableBeltDirectionComboBox ? TEXT("Valid") : TEXT("NULL"),
-        BeltTierToBuildingComboBox ? TEXT("Valid") : TEXT("NULL"),
-        BeltChainCheckBox ? TEXT("Valid") : TEXT("NULL"));
-
-    // Update Enabled checkbox
-    if (BeltEnabledCheckBox)
-    {
-        BeltEnabledCheckBox->SetIsChecked(Settings.bStackableBeltEnabled);
-    }
-
-    // Update Tier ComboBox (reuse main belt tier)
-    if (BeltTierMainComboBox)
-    {
-        int32 Index = CachedBeltTiers.Find(Settings.BeltTierMain);
-        if (Index == INDEX_NONE)
-        {
-            Index = 0;  // Default to Auto if current setting not in list
-        }
-        BeltTierMainComboBox->SetSelectedIndex(Index);
-    }
-
-    // Use dedicated StackableBeltDirectionComboBox for direction selection
-    if (StackableBeltDirectionComboBox)
-    {
-        StackableBeltDirectionComboBox->ClearOptions();
-        StackableBeltDirectionComboBox->AddOption(TEXT("Forward"));
-        StackableBeltDirectionComboBox->AddOption(TEXT("Backward"));
-
-        int32 DirectionIndex = Settings.StackableBeltDirection;
-        StackableBeltDirectionComboBox->SetSelectedIndex(DirectionIndex);
-    }
-
-    // Show direction row, hide irrelevant rows for stackable poles
-    if (StackableBeltDirectionRow)
-    {
-        StackableBeltDirectionRow->SetVisibility(ESlateVisibility::Visible);
-    }
-    if (BeltTierToBuildingComboBox)
-    {
-        BeltTierToBuildingComboBox->SetVisibility(ESlateVisibility::Collapsed);
-    }
-    if (BeltTierToBuildingLabel)
-    {
-        BeltTierToBuildingLabel->SetVisibility(ESlateVisibility::Collapsed);
-    }
-    if (BeltChainCheckBox)
-    {
-        BeltChainCheckBox->SetVisibility(ESlateVisibility::Collapsed);
-    }
-    if (BeltChainLabel)
-    {
-        BeltChainLabel->SetVisibility(ESlateVisibility::Collapsed);
-    }
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Updated stackable belt controls - Enabled=%d, Tier=%d, Direction=%d"),
-        Settings.bStackableBeltEnabled, Settings.BeltTierMain, Settings.StackableBeltDirection);
-}
-
-void USmartSettingsFormWidget::UpdateStackablePipeAutoConnectControls()
-{
-    if (!CachedSubsystem.IsValid())
-    {
-        return;
-    }
-
-    const auto& Settings = CachedSubsystem->GetAutoConnectRuntimeSettings();
-
-    // Update Enabled checkbox
-    if (PipeEnabledCheckBox)
-    {
-        PipeEnabledCheckBox->SetIsChecked(Settings.bPipeAutoConnectEnabled);
-    }
-
-    // Update To Building tier ComboBox (Stackable and Floor holes use this per user request)
-    if (PipeTierToBuildingComboBox)
-    {
-        int32 Index = CachedPipeTiers.Find(Settings.PipeTierToBuilding);
-        if (Index == INDEX_NONE)
-        {
-            Index = 0;  // Default to Auto if current setting not in list
-        }
-        PipeTierToBuildingComboBox->SetSelectedIndex(Index);
-    }
-
-    // Hide irrelevant controls for stackable pipe supports
-    if (PipeTierMainRow) PipeTierMainRow->SetVisibility(ESlateVisibility::Collapsed);
-    if (PipeTierToBuildingRow) PipeTierToBuildingRow->SetVisibility(ESlateVisibility::Visible);
-
-    // Update Flow Indicator checkbox and row visibility
-    // Only show if clean pipes are unlocked - otherwise flow indicators are implicit
-    AFGPlayerController* PC = Cast<AFGPlayerController>(GetOwningPlayer());
-    bool bCleanPipesUnlocked = CachedSubsystem->AreCleanPipesUnlocked(PC);
-
-    if (PipeIndicatorRow)
-    {
-        PipeIndicatorRow->SetVisibility(bCleanPipesUnlocked ? ESlateVisibility::Visible : ESlateVisibility::Collapsed);
-    }
-
-    if (PipeIndicatorCheckBox)
-    {
-        // If clean pipes not unlocked, force to true (with indicators)
-        if (!bCleanPipesUnlocked)
-        {
-            PipeIndicatorCheckBox->SetIsChecked(true);
-        }
-        else
-        {
-            PipeIndicatorCheckBox->SetIsChecked(Settings.bPipeIndicator);
-        }
-    }
-
-    // Update Routing Mode ComboBox (same as regular pipe controls)
-    if (PipeRoutingModeComboBox)
-    {
-        PipeRoutingModeComboBox->ClearOptions();
-        PipeRoutingModeComboBox->AddOption(TEXT("Auto"));
-        PipeRoutingModeComboBox->AddOption(TEXT("Auto 2D"));
-        PipeRoutingModeComboBox->AddOption(TEXT("Straight"));
-        PipeRoutingModeComboBox->AddOption(TEXT("Curve"));
-        PipeRoutingModeComboBox->AddOption(TEXT("Noodle"));
-        PipeRoutingModeComboBox->AddOption(TEXT("Horiz→Vert"));
-
-        int32 ModeIndex = FMath::Clamp(Settings.PipeRoutingMode, 0, 5);
-        PipeRoutingModeComboBox->SetSelectedIndex(ModeIndex);
-    }
-
-    UE_LOG(LogSmartFoundations, VeryVerbose, TEXT("Settings Form: Updated stackable/floor hole pipe controls - Enabled=%d, Tier=%d, Indicator=%d, RoutingMode=%d, CleanUnlocked=%d"),
-        Settings.bPipeAutoConnectEnabled, Settings.PipeTierToBuilding, Settings.bPipeIndicator, Settings.PipeRoutingMode, bCleanPipesUnlocked);
-}
-
 #undef LOCTEXT_NAMESPACE
